Fix parameter binding
Consider a client that wants to perform
 Invoke-CheckFoo -Bar baz,qux

via the API. It serializes the arguments list as
JSON {"Bar", "baz,qux"}, and HTTP POSTs it to
/v1/checker?command=Invoke-CheckFoo.

At the service, the former Invoke-Command usage ten expanded like this:
 Invoke-Command -ScriptBlock { return &"Invoke-CheckFoo" -Bar "baz,qux" };

The client ment to pass an array to parameter Bar, but the value
eventually is passed as string by the apicheck service.
A plugin receiving a different type than it expects is likely to fail.

Now we assemble a complete command line as it would look like
if called as CmdLet for evaluation by Invoke-Command. Then types
are handled as expected by the parameter binder.